system/templates/Script Tool Template/requirements.txt
# [Tool Name] Dependencies
# Install with: pip3 install -r requirements.txt

# Environment variable loading (required for /memory/ config)
python-dotenv>=1.0.0

# Add your dependencies below
# requests>=2.31.0
# pillow>=10.0.0


